Adapt to move of StringLiteral
