+++ score/mw/com/mocking/BUILD
@@ -33,8 +33,8 @@ cc_library(
         "//score/mw/com:types",
         "//score/mw/com/impl:instance_specifier",
         "@score_baselibs//score/language/futurecpp",
-        "@score_baselibs//score/memory:string_literal",
         "@score_baselibs//score/result",
+        "@score_baselibs//score/string_manipulation:string_literal",
     ],
 )
