+++ score/mw/com/mocking/i_runtime.h
@@ -17,8 +17,8 @@
 
 #include "score/mw/com/types.h"
 
-#include "score/memory/string_literal.h"
 #include "score/result/result.h"
+#include "score/string_manipulation/string_literal.h"
 
 #include <score/span.hpp>
