+++ score/mw/com/runtime.h
@@ -24,8 +24,8 @@
 #include "score/mw/com/runtime_configuration.h"
 #include "score/mw/com/types.h"
 
-#include "score/memory/string_literal.h"
 #include "score/result/result.h"
+#include "score/string_manipulation/string_literal.h"
 
 #include <cstdint>
