+++ score/mw/com/runtime_configuration_test.cpp
@@ -12,9 +12,9 @@
  ********************************************************************************/
 #include "score/mw/com/runtime_configuration.h"
 
-#include "score/memory/string_literal.h"
 #include "score/mw/log/logging.h"
 #include "score/mw/log/recorder_mock.h"
+#include "score/string_manipulation/string_literal.h"
 
 #include <score/assert.hpp>
 #include <score/span.hpp>
